requirements.txt
fastapi
httpx
openai
uvicorn
pydantic
sqlalchemy
alembic
dotenv
botbuilder.core
botbuilder.schema
python-dotenv
aiohttp